+++ crates/io/src/store.rs
@@ -1,7 +1,8 @@
 use crate::{Format, Readable, Result, Writeable};
 use object_store::{ObjectStore, ObjectStoreScheme, PutResult, path::Path};
 use stac::Href;
-use std::sync::Arc;
+use std::{fmt::Debug, sync::Arc};
+use tracing::instrument;
 use url::Url;
 
 /// Parses an href into a [StacStore] and a [Path].
@@ -72,7 +73,7 @@ where
 }
 
 /// Reads STAC from an [ObjectStore].
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct StacStore {
     store: Arc<dyn ObjectStore>,
     root: Url,
@@ -126,7 +127,8 @@ impl StacStore {
     }
 
     /// Gets a STAC value from the store in a specific format.
-    pub async fn get_format<T>(&self, path: impl Into<Path>, format: Format) -> Result<T>
+    #[instrument(skip(self))]
+    pub async fn get_format<T>(&self, path: impl Into<Path> + Debug, format: Format) -> Result<T>
     where
         T: Readable,
     {
@@ -141,7 +143,7 @@ impl StacStore {
     /// Puts a STAC value to the store.
     pub async fn put<T>(&self, path: impl Into<Path>, value: T) -> Result<PutResult>
     where
-        T: Writeable,
+        T: Writeable + Debug,
     {
         let path = path.into();
         let format = Format::infer_from_href(path.as_ref()).unwrap_or_default();
@@ -149,14 +151,15 @@ impl StacStore {
     }
 
     /// Puts a STAC value to the store in a specific format.
+    #[instrument(skip(self))]
     pub async fn put_format<T>(
         &self,
-        path: impl Into<Path>,
+        path: impl Into<Path> + Debug,
         value: T,
         format: Format,
     ) -> Result<PutResult>
     where
-        T: Writeable,
+        T: Writeable + Debug,
     {
         let path = path.into();
         let bytes = format.into_vec(value)?;